+++ src/main/kotlin/com/atvgstudios/klib/typealiases/Function.kt
@@ -7,3 +7,12 @@ package com.atvgstudios.klib.typealiases
  * @author Nils Rider
  */
 typealias Function = () -> Unit
+
+
+/**
+ * Type alias for () -> Any?
+ *
+ * @since 0.1.6
+ * @author Thomas Obernosterer
+ */
+typealias Function_Any = () -> Any?