+++ src/test/kotlin/klib/extensions/TAny.kt
@@ -1,24 +1,39 @@
 package klib.extensions
 
 import com.atvgstudios.klib.extensions.ofType
+import com.atvgstudios.klib.extensions.or
+import com.atvgstudios.klib.extensions.orFun
 import org.junit.Test
 
-open class Animal(val name: String)
+open class Animal(val name: String, val isAnimal: Boolean = true)
 
-class Dog(name: String) : Animal(name)
+class Dog(name: String, val isDog: Boolean = true) : Animal(name)
 
 class TAny {
     @Test
     fun ofType(){
         var success = false
 
-        var peter = Animal("Peter")
-        peter = Dog("Peter")
+        val peter: Animal = Dog("Peter")
 
         peter.ofType<Dog> {
-            success = it.name == "Peter"
+            success = it.isAnimal && it.isDog
         }
 
         assert(success)
+    }
+
+    @Test
+    fun or() {
+        val test: Any? = null
+        assert(test or true == true)
+    }
+
+    @Test
+    fun orFun() {
+        val test: Any? = null
+        assert(test orFun {
+            true
+        } == true)
     }
 }